refactor(cookie-consent): co-locate useConsent hook in provider (#2340)
